Fixed VC++ compilation:sqrt' : ambiguous call to overflow function (line 25)
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@9240 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/video/ZMBVEncoder.cc
@@ -22,7 +22,7 @@ static const unsigned FLAG_KEYFRAME = 0x01;
 
 struct CodecVector {
 	double cost() const {
-		double c = sqrt(x * x + y * y);
+		double c = sqrt(double(x * x + y * y));
 		if ((x == 0) || (y == 0)) {
 			// no penalty for purely horizontal/vertical offset
 			c *= 1.0;
